requirements.txt
streamlit
pandas
plotly
geopandas
fpdf2
kaleido
xlsxwriter

